# Simplify some term search tactics