… crates/ide-assists/src/handlers/term_search.rs
… crates/ide-completion/src/render.rs
… crates/rust-analyzer/src/config.rs
… docs/user/generated_config.adoc
… editors/code/package.json